py_crow_test/requirements.txt
RPi.GPIO==0.6.3
argparse==1.2.1
spidev==3.2
wsgiref==0.1.2
